Tooltip Styles and Options.
It's been a while since i worked on this project so here is my contribution:

I added tooltip background styles based on the tooltip styles provided by BTA and an option to reduce the opacity of the background.

i also changed the way inventory is displayed not item_list. they should now render only 3 rows of item until minimum size is reached.

Right now the textures loaded are the ones in minecraft/gui/.../xxx.png to make custom ones we'll have to change the string.